# src/main/resources/static/swagger.json
{"openapi":"3.1.0","info":{"title":"Book Catalog API","description":"REST API for managing a book catalog","contact":{"name":"Developer","email":"dev@example.com"},"version":"1.0.0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"tags":[{"name":"Books","description":"Book CRUD API"}],"paths":{"/api/books/{id}":{"get":{"tags":["Books"],"summary":"Get a book by ID","description":"Returns a single book by its unique identifier","operationId":"getById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"Book found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Book"}}}},"404":{"description":"Book not found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}}}},"put":{"tags":["Books"],"summary":"Update an existing book","description":"Updates the details of an existing book by ID","operationId":"update","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Book"}}},"required":true},"responses":{"200":{"description":"Book updated successfully","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Book"}}}},"404":{"description":"Book not found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}}}},"delete":{"tags":["Books"],"summary":"Delete a book","description":"Removes a book from the catalog by ID","operationId":"delete","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"204":{"description":"Book deleted successfully"},"404":{"description":"Book not found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}}}}},"/api/books":{"get":{"tags":["Books"],"summary":"List all books","description":"Returns a list of all books in the catalog","operationId":"getAll","responses":{"200":{"description":"List of books retrieved successfully","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Book"}}}}}}},"post":{"tags":["Books"],"summary":"Create a new book","description":"Adds a new book to the catalog","operationId":"create","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Book"}}},"required":true},"responses":{"201":{"description":"Book created successfully","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Book"}}}},"400":{"description":"Invalid input","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}}}}},"/api/books/paged":{"get":{"tags":["Books"],"summary":"List books with pagination","description":"Returns a paginated list of books","operationId":"getAllPaged","parameters":[{"name":"page","in":"query","description":"Page number (0-based)","required":false,"schema":{"type":"integer","format":"int32","default":0,"description":"Page number (0-based)","example":0},"example":0},{"name":"size","in":"query","description":"Number of items per page","required":false,"schema":{"type":"integer","format":"int32","default":10,"description":"Number of items per page","example":10},"example":10}],"responses":{"200":{"description":"Paginated list of books retrieved successfully","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ServiceResponsePagedResponseBook"}}}}}}}},"components":{"schemas":{"Book":{"type":"object","description":"Represents a book in the catalog","properties":{"id":{"type":"integer","format":"int64","description":"Unique identifier","example":1,"readOnly":true},"title":{"type":"string","description":"Book title","example":"The Hobbit","minLength":1},"author":{"type":"string","description":"Author name","example":"J.R.R. Tolkien","minLength":1},"price":{"type":"number","description":"Price in USD","example":"29.99"},"category":{"type":"string","description":"Book category/genre","example":"Fantasy"},"createdAt":{"type":"string","format":"date-time","description":"Timestamp when the book was added","example":"2026-05-12T18:00:00","readOnly":true}},"required":["author","price","title"]},"ProblemDetail":{"type":"object","properties":{"type":{"type":"string","format":"uri"},"title":{"type":"string"},"status":{"type":"integer","format":"int32"},"detail":{"type":"string"},"instance":{"type":"string","format":"uri"},"properties":{"type":"object","additionalProperties":{}}}},"PagedResponseBook":{"type":"object","description":"Paginated response wrapper","properties":{"page":{"type":"integer","format":"int32","description":"Current page number (0-based)","example":0},"size":{"type":"integer","format":"int32","description":"Number of items per page","example":10},"totalElements":{"type":"integer","format":"int64","description":"Total number of items across all pages","example":42},"content":{"type":"array","description":"Page content","items":{"$ref":"#/components/schemas/Book"}}}},"ServiceResponsePagedResponseBook":{"type":"object","description":"Generic API response wrapper","properties":{"success":{"type":"boolean","description":"Indicates if the request was successful","example":true},"message":{"type":"string","description":"Response message","example":"Books retrieved successfully"},"data":{"$ref":"#/components/schemas/PagedResponseBook","description":"Response payload"}}}}}}